Move test for reflection destructurer to the ReflectionBasedDestructurerTest class
